Allow all collections of spots to include full spot detail
Allow all client queries that return a collection of spots to specify
the `detail: true` option, which will return the full spot detail for
every spot in the collection.

This option should be used with care, as it adds an additional API call
for EACH spot in the collection. E.g. a spots collection of 100 spots
will use 101 API calls when the `detail: true` option is set.